removed brackets

+++ tests/Get-DbaAvailabilityGroup.Tests.ps1
@@ -23,6 +23,7 @@ Describe "$commandname Unit Tests" -Tag 'UnitTests' {
 
 Describe "$commandname Integration Tests" -Tag "IntegrationTests" {
     BeforeAll {
+        # $script:instance2 - to make it appear in the proper place on appveyor
         Get-DbaProcess -SqlInstance $script:instance3 -Program 'dbatools PowerShell module - dbatools.io' | Stop-DbaProcess -WarningAction SilentlyContinue
         $server = Connect-DbaInstance -SqlInstance $script:instance3
         $computername = $server.NetName
@@ -30,22 +31,22 @@ Describe "$commandname Integration Tests" -Tag "IntegrationTests" {
         $server.Query("create database $dbname")
         $backup = Get-DbaDatabase -SqlInstance $script:instance3 -Database $dbname | Backup-DbaDatabase
         $server.Query("IF NOT EXISTS (select * from sys.symmetric_keys where name like '%DatabaseMasterKey%') CREATE MASTER KEY ENCRYPTION BY PASSWORD = '<StrongPassword>'")
-        $server.Query("IF EXISTS ( SELECT * FROM sys.tcp_endpoints WHERE name = 'End_Mirroring') DROP ENDPOINT [endpoint_mirroring]")
+        $server.Query("IF EXISTS ( SELECT * FROM sys.tcp_endpoints WHERE name = 'End_Mirroring') DROP ENDPOINT endpoint_mirroring")
         $server.Query("CREATE CERTIFICATE dbatoolsci_AGCert WITH SUBJECT = 'AG Certificate'")
         $server.Query("CREATE ENDPOINT dbatoolsci_AGEndpoint
                             STATE = STARTED
                             AS TCP (LISTENER_PORT = 5022,LISTENER_IP = ALL)
                             FOR DATABASE_MIRRORING (AUTHENTICATION = CERTIFICATE dbatoolsci_AGCert,ROLE = ALL)")
-        $server.Query("CREATE AVAILABILITY GROUP [dbatoolsci_agroup]
+        $server.Query("CREATE AVAILABILITY GROUP dbatoolsci_agroup
                             WITH (DB_FAILOVER = OFF, DTC_SUPPORT = NONE, CLUSTER_TYPE = NONE)
-                            FOR DATABASE [$dbname] REPLICA ON N'$script:instance3' 
+                            FOR DATABASE $dbname REPLICA ON N'$script:instance3'
                             WITH (ENDPOINT_URL = N'TCP://$computername`:5022', FAILOVER_MODE = MANUAL, AVAILABILITY_MODE = SYNCHRONOUS_COMMIT)")
     }
     AfterAll {
-        Remove-Item -Path $backup.BackupPath -ErrorAction SilentlyContinue
-        $server.Query("DROP AVAILABILITY GROUP [dbatoolsci_agroup]")
+        if ($backup.BackupPath) { Remove-Item -Path $backup.BackupPath -ErrorAction SilentlyContinue }
+        $server.Query("DROP AVAILABILITY GROUP dbatoolsci_agroup")
         Get-DbaDatabase -SqlInstance $script:instance3 -Database $dbname | Remove-DbaDatabase -Confirm:$false
-        $server.Query("DROP ENDPOINT [dbatoolsci_AGEndpoint]")
+        $server.Query("DROP ENDPOINT dbatoolsci_AGEndpoint")
         $server.Query("DROP CERTIFICATE dbatoolsci_AGCert")
     }
